+++ client/components/admin-page-orders.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import {Container, Card} from 'semantic-ui-react'
+
+const AdminOrderPage = () => {
+  return (
+    <Container>
+      <h1>view all orders here</h1>
+    </Container>
+  )
+}
+
+export default AdminOrderPage
