+++ client/components/admin-page-users.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import {Container, Card} from 'semantic-ui-react'
+
+const AdminUserPage = () => {
+  return (
+    <Container>
+      <h1>view all users here</h1>
+    </Container>
+  )
+}
+
+export default AdminUserPage
