+++ client/components/index.js
@@ -7,7 +7,9 @@ export {default as Navbar} from './navbar'
 export {default as UserHome} from './user-home'
 export {Login, Signup} from './auth-form'
 export {default as AdminPage} from './admin-page'
-export {default as AdminForm} from './admin-form'
+export {default as AdminHomePage} from './admin-home-page'
+export {default as AdminOrderPage} from './admin-page-orders'
+export {default as AdminUserPage} from './admin-page-users'
 export {default as DetailedBook} from './detailed-book'
 export {default as Books} from './books'
 export {default as Search} from './search'
